+++ test-plugins/org.yakindu.sct.model.sexec.test/src/org/yakindu/sct/model/sexec/transformation/test/Assert.java
@@ -91,7 +91,7 @@ public class Assert {
 						.getValue());
 			}
 			else if (literal instanceof RealLiteral) {
-				currentValue = Float.toString(((RealLiteral) literal)
+				currentValue = Double.toString(((RealLiteral) literal)
 						.getValue());
 			}
 			else if (literal instanceof StringLiteral) {
